fix: update exports
fix: update exports

fix: update imports

fix: update export

fix: update imports

fix: update export

fix: update type

chore: creat services folder

delete pinned services context

Delete index.ts

update import

Update index.tsx

Update PinnedGrid.tsx

Update ServicesGrid.tsx

Update ServicesGrid.tsx

Revert "Revert "Update PinnedDialog.tsx""

This reverts commit 7ecc85be9efdb458f1fadf746cb20ca92d2d8462.

Update index.tsx

Revert "Update PinnedDialog.tsx"

This reverts commit bed2ce1c4eca276d0d25b1be74fa2e1632c888b8.

Update PinnedDialog.tsx

Update PinnedDialog.tsx

Update PinnedDialog.tsx
