+++ src/context/categories/CategoriesContext.ts
@@ -1,10 +1,10 @@
 import React, { createContext } from "react";
 import { Category } from "../../interfaces";
 
-interface CategoriesContextProps {
+type CategoriesContextProps = {
   categories: Category[];
   setCategories: React.Dispatch<React.SetStateAction<Category[]>>;
-}
+};
 
 export const CategoriesContext = createContext<CategoriesContextProps>({
   categories: [],
