+++ src/context/categories/index.ts
@@ -1,0 +1,3 @@
+export { CategoriesContext } from "./CategoriesContext";
+export { CategoriesProvider } from "./CategoriesProvider";
+export { useCategory } from "./useCategory";
