+++ src/context/categories/useCategory.ts
@@ -1,0 +1,4 @@
+import { useContext } from "react";
+import { CategoriesContext } from "./CategoriesContext";
+
+export const useCategory = () => useContext(CategoriesContext);
