+++ src/hooks/usePinnedServices.ts
@@ -1,11 +1,6 @@
 import { Service } from "../interfaces";
 
 const usePinnedServices = () => {
-  const fetchPinnedServices = async () => {
-    const response = await fetch("/api/pinned-services", { method: "GET" });
-    return await response.json();
-  };
-
   const addPinnedService = async (services: Service[]) => {
     try {
       const response = await fetch("/api/update-pinned", {
@@ -22,7 +17,7 @@ const usePinnedServices = () => {
     }
   };
 
-  return { fetchPinnedServices, addPinnedService };
+  return { addPinnedService };
 };
 
 export default usePinnedServices;
