+++ src/services/fetchPinnedServices.ts
@@ -1,0 +1,6 @@
+const fetchPinnedServices = async () => {
+  const response = await fetch("/api/pinned-services", { method: "GET" });
+  return await response.json();
+};
+
+export default fetchPinnedServices;
