+++ src/services/index.ts
@@ -1,0 +1,4 @@
+import fetchCategories from "./fetchCategories";
+import fetchPinnedServices from "./fetchPinnedServices";
+
+export { fetchCategories, fetchPinnedServices };
